Fixes for Issues #3, #5, and #6
Issue #3: Added media query to adjust the position of the background
image and changed the layout to a 2 row format for the jumbotron.
Issue #5: Adjusted JS for chart generation to modify label placement,
modified content of the CSV label, and altered CSS to set a min and max
height for the chart area on all screen sizes.
Issue #6: Removed extra '>' character.